ENH: clean buildsystem
There were a lot of ifles that were not really needed to make this PDF
compile. Further, it was rather hard to understand what was the entry
point given the "generic" nature of the repository.

Remove those files, simplify the makefile (to use latexmk) and
change the input file to something more descriptive (diffbom).