ENH: split content into smaller files
The current setup was akin  to having a main() function with everything
in it. We can do better.

Instead, split the paper into file-per-section, to simplify re-arrangement,
minimize git conflicts and help us separate things conceptually a little
bit better. Likewise, rename ccs-body to diffbom-body to keep things
clearer. Although ccs-body may not be needed in the way it stands (it
could easily be inlined into diffbom.tex later), we will leave that
question for a future enhancement.